chore: misc file managment to clean up current work directory
